tax-agent-backend/requirements.txt
# requirements.txt
langchain
openai
pymupdf
python-dotenv 
fastapi
uvicorn 